requirements.txt
pyyaml
pandas
numpy
matplotlib
os-sys
importlib
argparse
requires.io
inspect.it
plotly
multiprocess
poli-sci-kit